+++ tokio-util/src/io/mod.rs
@@ -10,6 +10,7 @@
 //! [`Body`]: https://docs.rs/hyper/0.13/hyper/struct.Body.html
 //! [`AsyncRead`]: tokio::io::AsyncRead
 
+mod copy_to_bytes;
 mod inspect;
 mod read_buf;
 mod reader_stream;
@@ -21,9 +22,10 @@ cfg_io_util! {
     pub use self::sync_bridge::SyncIoBridge;
 }
 
+pub use self::copy_to_bytes::CopyToBytes;
 pub use self::inspect::{InspectReader, InspectWriter};
 pub use self::read_buf::read_buf;
 pub use self::reader_stream::ReaderStream;
-pub use self::sink_writer::{CopyToBytes, SinkWriter};
+pub use self::sink_writer::SinkWriter;
 pub use self::stream_reader::StreamReader;
 pub use crate::util::{poll_read_buf, poll_write_buf};
